requirements.txt
aiogram=>3.20.0.post0
sqlalchemy=>2.0.41
aiosqlite=>0.21.0
apscheduler=>3.11.0
python-dotenv=>1.1.0